fix: フォルダ配置の修正と関数コンポーネントへの修正
