+++ src/app/history/utils.ts
@@ -6,17 +6,3 @@ export type Day = {
   colors: { r: number; g: number; b: number }[]
   pattern: number // 0 <= pattern < 5
 }
-
-export type FridgeViewProps = {
-  historyData: Day[]
-  onIceCellTap: (element: Day) => void
-}
-
-export type IceCellProps = {
-  iceCell: Day
-  onIceCellTap: (element: Day) => void
-}
-
-export type PhotoViewProps = {
-  selectedElement: Day | null
-}
