+++ src/layouts/Fridge/FridgeView.module.css
