+++ src/layouts/Fridge/IceCell/IceCell.module.css
